refactor: move around more things to appropriate places
