+++ bin/sozo/src/commands/mod.rs
@@ -1,7 +1,6 @@
 use anyhow::Result;
+use clap::{command, Subcommand};
 use scarb::core::Config;
-
-use crate::args::Commands;
 
 pub(crate) mod auth;
 pub(crate) mod build;
@@ -17,6 +16,51 @@ pub(crate) mod model;
 pub(crate) mod options;
 pub(crate) mod register;
 pub(crate) mod test;
+
+use auth::AuthArgs;
+use build::BuildArgs;
+use call::CallArgs;
+use clean::CleanArgs;
+use completions::CompletionsArgs;
+use dev::DevArgs;
+use events::EventsArgs;
+use execute::ExecuteArgs;
+use init::InitArgs;
+use migrate::MigrateArgs;
+use model::ModelArgs;
+use register::RegisterArgs;
+use test::TestArgs;
+
+#[derive(Subcommand)]
+pub enum Commands {
+    #[command(about = "Build the world, generating the necessary artifacts for deployment")]
+    Build(BuildArgs),
+    #[command(about = "Initialize a new project")]
+    Init(InitArgs),
+    #[command(about = "Remove generated artifacts, manifests and abis")]
+    Clean(CleanArgs),
+    #[command(about = "Run a migration, declaring and deploying contracts as necessary to \
+                       update the world")]
+    Migrate(Box<MigrateArgs>),
+    #[command(about = "Developer mode: watcher for building and migration")]
+    Dev(DevArgs),
+    #[command(about = "Test the project's smart contracts")]
+    Test(TestArgs),
+    #[command(about = "Execute a world's system")]
+    Execute(ExecuteArgs),
+    #[command(about = "Call a world's system")]
+    Call(CallArgs),
+    #[command(about = "Interact with a worlds models")]
+    Model(ModelArgs),
+    #[command(about = "Register new models")]
+    Register(RegisterArgs),
+    #[command(about = "Queries world events")]
+    Events(EventsArgs),
+    #[command(about = "Manage world authorization")]
+    Auth(AuthArgs),
+    #[command(about = "Generate shell completion file for specified shell")]
+    Completions(CompletionsArgs),
+}
 
 pub fn run(command: Commands, config: &Config) -> Result<()> {
     match command {
